\name{annotation}
\docType{data}
\alias{annotation}
\title{QSEA example annotation}
\description{
  This is a list of GRange objects, which contain genomic annotations for hg19 reference, 
  obtained from UCSC table browser. 
}
\usage{ data(annotation) }
\examples{
\dontrun{
data(annotation)
library(GenomicRanges)
names(ROIs)
ROIs$'gene body'
names(tfbs)
}
}
\keyword{datasets}
