\name{CS}
\docType{data}
\alias{CS}
\title{COUPLING SET}
\description{
  This is a CpG coupling set generated by the MEDIPS package based on the human chromosome 22 (hg19) and with a window size of 100bp.
}
\usage{ data(CS) }
\examples{
\dontrun{
data(CS)
library(MEDIPS)
CS
}
}
\keyword{datasets}
