\name{mart_gene}
\docType{data}
\alias{mart_gene}
\title{An annotation object generated by accessing biomaRt using the MEDIPS.getAnnotation function of the MEDIPS package.}
\description{
This is an annotation object generated by accessing biomaRt using the MEDIPS.getAnnotation function of the MEDIPS package:
mart_gene = MEDIPS.getAnnotation(mart="ensembl", dataset=c("hsapiens_gene_ensembl"), annotation=c("GENE"), chr=22)
The annotation object contains genomic coordinates of human genes on chromosome 22.
}
\usage{ data(mart_gene) }
\examples{
\dontrun{
data(mart_gene)
data(resultTable)
library(MEDIPS)
resultTable = MEDIPS.setAnnotation(regions=resultTable, annotation=mart_gene)
}
}
\keyword{datasets}
