\name{samplesNSCLC}
\docType{data}
\alias{samplesNSCLC}
\title{QSEA lung cancer MeDIP seq sample table}
\description{
  This data set contains a sample table describing the samples of the 
    QSEA lung cancer MeDIP seq example data. 
}
\usage{ data(samplesNSCLC) }
\examples{
\dontrun{
data(samplesNSCLC)
samplesNSCLC
}
}
\keyword{datasets}
