\name{tcga_luad_lusc_450kmeth}
\docType{data}
\alias{tcga_luad_lusc_450kmeth}
\title{Lung cancer calibration data}
\description{
  Calibration data for the lung cancer MeDIP seq example data, taken from TCGA LUAD and LUSC studies.
  
}
\usage{ data(CS) }
\examples{
\dontrun{
data(tcga_luad_lusc_450kmeth)
tcga_luad_lusc_450kmeth
}
}
\keyword{datasets}
